+++ Controllers/SupplierController.cs
@@ -43,7 +43,7 @@ namespace FurnitureERP.Controllers
         }
 
         [Authorize]
-        public static async Task<IResult> GetSuppItems(AppDbContext db, IMapper mapper, int id, HttpRequest request)
+        public static async Task<IResult> GetSuppItems(AppDbContext db, IMapper mapper, long id, HttpRequest request)
         {
             if (!await db.Supps.AnyAsync(x => x.Id == id && x.MerchantGuid == request.GetCurrentUser().MerchantGuid))
             {
@@ -120,14 +120,14 @@ namespace FurnitureERP.Controllers
 
 
         [Authorize]
-        public static async Task<IResult> Single(AppDbContext db, int id, IMapper mapper, HttpRequest request)
+        public static async Task<IResult> Single(AppDbContext db, long id, IMapper mapper, HttpRequest request)
         {
             var et = await db.Supps.SingleOrDefaultAsync(x => x.Id == id && x.MerchantGuid == request.GetCurrentUser().MerchantGuid);
             return et == null ? Results.NotFound() : Results.Ok(mapper.Map<SuppDto>(et));
         }
 
         [Authorize]
-        public static async Task<IResult> Edit(AppDbContext db, int id, CreateSuppDto SuppDto, HttpRequest request, IMapper mapper)
+        public static async Task<IResult> Edit(AppDbContext db, long id, CreateSuppDto SuppDto, HttpRequest request, IMapper mapper)
         {
             var UserName = request.GetCurrentUser().UserName;
             var MerchantGuid = request.GetCurrentUser().MerchantGuid;
@@ -167,7 +167,7 @@ namespace FurnitureERP.Controllers
         }
 
         [Authorize]
-        public static async Task<IResult> Delete(AppDbContext db, int id, HttpRequest request)
+        public static async Task<IResult> Delete(AppDbContext db, long id, HttpRequest request)
         {
             var et = await db.Supps.FirstOrDefaultAsync(x => x.Id == id && x.MerchantGuid == request.GetCurrentUser().MerchantGuid);
             if (et == null)
